refactor: move driver trait to sub module
